move overlaydb to ref_overlaydb, add new overlaydb without on-disk rc
