+++ util/src/journaldb/refcounteddb.rs
@@ -19,7 +19,7 @@
 use common::*;
 use rlp::*;
 use hashdb::*;
-use overlaydb::*;
+use ref_overlaydb::RefOverlayDB;
 use super::{DB_PREFIX_LEN, LATEST_ERA_KEY, VERSION_KEY};
 use super::traits::JournalDB;
 use kvdb::{Database, DBTransaction, DatabaseConfig};
@@ -34,7 +34,7 @@ use std::env;
 /// immediately. Rather some age (based on a linear but arbitrary metric) must pass before
 /// the removals actually take effect.
 pub struct RefCountedDB {
-	forward: OverlayDB,
+	forward: RefOverlayDB,
 	backing: Arc<Database>,
 	latest_era: Option<u64>,
 	inserts: Vec<H256>,
@@ -63,7 +63,7 @@ impl RefCountedDB {
 		let latest_era = backing.get(&LATEST_ERA_KEY).expect("Low-level database error.").map(|val| decode::<u64>(&val));
 
 		RefCountedDB {
-			forward: OverlayDB::new_with_arc(backing.clone()),
+			forward: RefOverlayDB::new_with_arc(backing.clone()),
 			backing: backing,
 			inserts: vec![],
 			removes: vec![],
