+++ util/src/lib.rs
@@ -138,6 +138,7 @@ pub mod hashdb;
 pub mod memorydb;
 pub mod migration;
 pub mod overlaydb;
+pub mod ref_overlaydb;
 pub mod journaldb;
 pub mod kvdb;
 mod math;
